<refactor>: refactor intro page
- change diretory
- rename files
- seperate features
