+++ src/features/scene/intro/hooks/use-animated-text.tsx
@@ -8,11 +8,11 @@ export function useAnimatedText(target, transition) {
   useEffect(() => {
     ref.current.innerText = target;
 
-    return value.on("change", (v) => {
+    return value.on("change", v => {
       ref.current.innerText = v;
     });
   });
-  useEffect(() => value.set(target), [target]);
+  useEffect(() => value.set(target), [target, value]);
 
   return ref;
 }
