+++ src/features/scene/intro/hooks/use-windowsize.tsx
@@ -1,0 +1,13 @@
+import { useEffect, useState } from "react";
+
+export const useWindowSize = () => {
+  const [size, setSize] = useState({ width: 0, height: 0 });
+  useEffect(() => {
+    const updateSize = () =>
+      setSize({ width: window.innerWidth, height: window.innerHeight });
+    updateSize();
+    window.addEventListener("resize", updateSize);
+    return () => window.removeEventListener("resize", updateSize);
+  }, []);
+  return size;
+};
